requirements.txt
binance==0.3
ipython==8.4.0
ipywidgets==7.6.5
numpy==1.21.5
pandas==1.4.2
plotly==5.7.0
python-dotenv==0.20.0
python_binance==1.0.16
python_dateutil==2.8.2
pytz==2022.1
requests==2.26.0
ruamel.base==1.0.0
